move blocked actions for modal
