+++ common/types/modal-requests.ts
@@ -1,5 +1,6 @@
 import {CardComponent} from '../components'
 import {CardEntity, PlayerEntity} from '../entities'
+import {TurnAction} from './game-state'
 
 export type ModalRequest =
 	| SelectCards.Request
@@ -118,6 +119,8 @@ export namespace CopyAttack {
 		hermitCard: CardEntity
 		/** Show a close button on this modal. */
 		cancelable: boolean
+		/** The actions that can not be used in this modal */
+		blockedActions: Array<TurnAction>
 	}
 
 	export type Result =
